Fixed level not being overridden correctly for "stored" logger
Changed the architecture of the v2 core a bit so that level can be correctly set even when logger are stored by structure.

When a logger is stored in a structure, the origin pointer address is copied. When the override was happening before, we would actually change the `zlog` variable to make it point to somewhere else. This means all previously copied address actually points to the old logger with the old level.

This was ultimately preventing those from showing up for example as debug level correctly. Now, when using the ApplicationLogger, all loggers are right away instantiated to their correct *zap.Logger instance. Later, when level needs to be switched, we simply update the atomic level of the logger (that we now keep alongside the logger).

This change also has the impact that any `PackageLogger` that arrive after the `ApplicationLogger` has been invoked will no be tied to the correct logger. That might be enforced in the future.